Refactor and split into functions start.
